Extracting State and Refactoring
