+++ components/math/multiplicationConnect/GameBoardBlock.tsx
@@ -2,7 +2,7 @@ import React, {useState} from 'react'
 
 
 const blockClick = (block) => {
-    (block.isSelected===false) ? block.isSelected=true: block.isSelected=false;
+    (block.isSelected===false) ? block.isSelected=true : block.isSelected=false;
     console.log(block);
     return {...block};
 }
@@ -13,7 +13,8 @@ const GameBoardBlock = ({ blockData, onClick }) => {
   return (
     <div
       onClick={()=>setBlock(blockClick(block))}
-      className={`flex justify-center items-center h-full w-full cursor-pointer rounded-full transition ${block.isSelected===false ? "hover:bg-[#F20000]/40" : "bg-[#F20000]/40"}`}
+      className={`flex justify-center items-center h-full w-full cursor-pointer rounded-full duration-300 
+            ${block.isSelected===false ? "hover:bg-[#F20000]/40 hover:animate-pulse" : "bg-[#F20000]/40"}`}
     >
       <p>{block.gridNumber}</p>
     </div>
